+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/mappers/MockServerHttpResponseToFullHttpResponse.java
@@ -39,7 +39,7 @@ public class MockServerHttpResponseToFullHttpResponse {
   }
 
   private HttpResponseStatus getStatus(HttpResponse httpResponse) {
-    int statusCode = httpResponse.getStatusCode() != null ? httpResponse.getStatusCode() : 200;
+    final int statusCode = httpResponse.getStatusCode() != null ? httpResponse.getStatusCode() : 200;
     final String reasonPhrase =
         httpResponse.getReasonPhrase() != null ? httpResponse.getReasonPhrase() : "";
     return new HttpResponseStatus(statusCode, reasonPhrase);