+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/mock/action/http/HttpActionHandler.java
@@ -22,8 +22,8 @@ import de.gematik.test.tiger.mockserver.mock.Expectation;
 import de.gematik.test.tiger.mockserver.mock.HttpAction;
 import de.gematik.test.tiger.mockserver.mock.HttpState;
 import de.gematik.test.tiger.mockserver.model.*;
+import de.gematik.test.tiger.mockserver.netty.responsewriter.NettyResponseWriter;
 import de.gematik.test.tiger.mockserver.proxyconfiguration.ProxyConfiguration;
-import de.gematik.test.tiger.mockserver.responsewriter.ResponseWriter;
 import de.gematik.test.tiger.mockserver.scheduler.Scheduler;
 import de.gematik.test.tiger.mockserver.socket.tls.NettySslContextFactory;
 import io.netty.buffer.Unpooled;
@@ -77,7 +77,7 @@ public class HttpActionHandler {
 
   public void processAction(
       final HttpRequest request,
-      final ResponseWriter responseWriter,
+      final NettyResponseWriter responseWriter,
       final ChannelHandlerContext ctx,
       Set<String> localAddresses,
       boolean proxyingRequest,
@@ -140,7 +140,7 @@ public class HttpActionHandler {
                       "Basic realm=\""
                           + StringEscapeUtils.escapeJava(configuration.proxyAuthenticationRealm())
                           + "\", charset=\"UTF-8\"");
-          responseWriter.writeResponse(request, response, false);
+          responseWriter.writeResponse(request, response);
           log.debug(
               "proxy authentication failed so returning response:{}for forwarded" + " request:{}",
               response,
@@ -191,7 +191,7 @@ public class HttpActionHandler {
                         request,
                         response);
                   }
-                  responseWriter.writeResponse(request, response, false);
+                  responseWriter.writeResponse(request, response);
                 } catch (SocketCommunicationException sce) {
                   log.warn("Exception while writing response", sce);
                   returnNotFound(responseWriter, request, sce.getMessage());
@@ -256,7 +256,7 @@ public class HttpActionHandler {
 
   public void writeForwardActionResponse(
       final HttpForwardActionResult responseFuture,
-      final ResponseWriter responseWriter,
+      final NettyResponseWriter responseWriter,
       final HttpRequest request,
       final Action action,
       boolean synchronous) {
@@ -268,7 +268,7 @@ public class HttpActionHandler {
                 responseFuture
                     .getHttpResponse()
                     .get(configuration.maxFutureTimeoutInMillis(), MILLISECONDS);
-            responseWriter.writeResponse(request, response, false);
+            responseWriter.writeResponse(request, response);
             log.debug(
                 "returning response:{}for forwarded request"
                     + NEW_LINE
@@ -296,9 +296,9 @@ public class HttpActionHandler {
   }
 
   public void writeForwardActionResponse(
-      final HttpResponse response, final ResponseWriter responseWriter, final HttpRequest request) {
+      final HttpResponse response, final NettyResponseWriter responseWriter, final HttpRequest request) {
     try {
-      responseWriter.writeResponse(request, response, false);
+      responseWriter.writeResponse(request, response);
       log.debug(
           "returning response:{}for forwarded request" + NEW_LINE + NEW_LINE + " in json:{}",
           response,
@@ -309,29 +309,34 @@ public class HttpActionHandler {
   }
 
   public void handleExceptionDuringForwardingRequest(
-      Action action, HttpRequest request, ResponseWriter responseWriter, Throwable exception) {
-    if (connectionException(exception)) {
-      log.error(
-          "failed to connect to remote socket while forwarding request {} for action {}",
-          request,
-          action,
-          exception);
-      returnNotFound(
-          responseWriter, request, "failed to connect to remote socket while forwarding request");
-    } else if (sslHandshakeException(exception)) {
-      log.error(
-          "TLS handshake exception while forwarding request {} for action {}",
-          request,
-          action,
-          exception);
-      returnNotFound(responseWriter, request, "TLS handshake exception while forwarding request");
+      Action action, HttpRequest request, NettyResponseWriter responseWriter, Throwable exception) {
+    if (action instanceof CloseChannel) {
+      log.debug("closing channel due to close action");
+      responseWriter.closeChannel();
     } else {
-      log.error("Exception while forwading request", exception);
-      returnNotFound(responseWriter, request, exception != null ? exception.getMessage() : null);
+      if (connectionException(exception)) {
+        log.error(
+            "failed to connect to remote socket while forwarding request {} for action {}",
+            request,
+            action,
+            exception);
+        returnNotFound(
+            responseWriter, request, "failed to connect to remote socket while forwarding request");
+      } else if (sslHandshakeException(exception)) {
+        log.error(
+            "TLS handshake exception while forwarding request {} for action {}",
+            request,
+            action,
+            exception);
+        returnNotFound(responseWriter, request, "TLS handshake exception while forwarding request");
+      } else {
+        log.error("Exception while forwarding request", exception);
+        returnNotFound(responseWriter, request, exception != null ? exception.getMessage() : null);
+      }
     }
   }
 
-  private void returnNotFound(ResponseWriter responseWriter, HttpRequest request, String error) {
+  private void returnNotFound(NettyResponseWriter responseWriter, HttpRequest request, String error) {
     HttpResponse response = notFoundResponse();
     if (request.getHeaders() != null
         && request
@@ -348,7 +353,7 @@ public class HttpActionHandler {
     } else {
       log.debug("no expectation for:{}returning response:{}", request, notFoundResponse());
     }
-    responseWriter.writeResponse(request, response, false);
+    responseWriter.writeResponse(request, response);
   }
 
   public HttpForwardActionHandler getHttpForwardActionHandler() {