+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/model/CloseChannel.java
@@ -1,0 +1,12 @@
+/*
+ * ${GEMATIK_COPYRIGHT_STATEMENT}
+ */
+
+package de.gematik.test.tiger.mockserver.model;
+
+/**
+ * Signifies a graceful channel close.
+ */
+public class CloseChannel extends Action {
+
+}