+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/netty/responsewriter/NettyResponseWriter.java
@@ -4,11 +4,15 @@
 
 package de.gematik.test.tiger.mockserver.netty.responsewriter;
 
+import static de.gematik.test.tiger.mockserver.model.HttpResponse.notFoundResponse;
+import static de.gematik.test.tiger.mockserver.model.HttpResponse.response;
+import static io.netty.handler.codec.http.HttpHeaderNames.CONTENT_LENGTH;
+import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
 import de.gematik.test.tiger.mockserver.configuration.Configuration;
 import de.gematik.test.tiger.mockserver.model.HttpRequest;
 import de.gematik.test.tiger.mockserver.model.HttpResponse;
-import de.gematik.test.tiger.mockserver.responsewriter.ResponseWriter;
+import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
 import io.netty.channel.ChannelFutureListener;
 import io.netty.channel.ChannelHandlerContext;
@@ -18,18 +22,46 @@ import lombok.extern.slf4j.Slf4j;
  * @author jamesdbloom
  */
 @Slf4j
-public class NettyResponseWriter extends ResponseWriter {
+public class NettyResponseWriter {
 
   private final ChannelHandlerContext ctx;
+  private final Configuration configuration;
 
   public NettyResponseWriter(
-      Configuration configuration,
-      ChannelHandlerContext ctx) {
-    super(configuration);
+    Configuration configuration,
+    ChannelHandlerContext ctx) {
+    this.configuration = configuration;
     this.ctx = ctx;
   }
 
-  @Override
+  public void writeResponse(final HttpRequest request, HttpResponse response) {
+    if (response == null) {
+      response = notFoundResponse();
+    }
+    String contentLengthHeader = response.getFirstHeader(CONTENT_LENGTH.toString());
+    if (isNotBlank(contentLengthHeader)) {
+      try {
+        int contentLength = Integer.parseInt(contentLengthHeader);
+        if (response.getBodyAsRawBytes().length > contentLength) {
+          log.info(
+            "returning response with content-length header {} which is smaller then response body length {}, body will likely be truncated by client receiving request",
+            contentLength,
+            response.getBodyAsRawBytes().length);
+        }
+      } catch (NumberFormatException ignore) {
+        log.trace("NumberFormatException while parsing content-length header", ignore);
+        // ignore exception while parsing invalid content-length header
+      }
+    }
+
+    // send response down the request HTTP2 stream
+    if (request.getStreamId() != null) {
+      response.withStreamId(request.getStreamId());
+    }
+
+    sendResponse(request, response);
+  }
+
   public void sendResponse(HttpRequest request, HttpResponse response) {
     writeAndCloseSocket(ctx, request, response);
   }
@@ -46,31 +78,38 @@ public class NettyResponseWriter extends ResponseWriter {
     }
   }
 
-  private void disconnectAndCloseChannel(ChannelFuture future) {
-    future
-        .channel()
-        .disconnect()
-        .addListener(
-            disconnectFuture -> {
-              if (disconnectFuture.isSuccess()) {
-                future
-                    .channel()
-                    .close()
-                    .addListener(
-                        closeFuture -> {
-                          if (disconnectFuture.isSuccess()) {
-                            if (log.isTraceEnabled()) {
-                              log.trace("disconnected and closed socket {}", future.channel().localAddress());
-                            }
-                          } else {
-                            if (log.isWarnEnabled()) {
-                              log.warn("exception closing socket {}", future.channel().localAddress(), disconnectFuture.cause());
-                            }
-                          }
-                        });
-              } else if (log.isWarnEnabled()) {
-                log.warn("exception disconnecting socket {}", future.channel().localAddress(), disconnectFuture.cause());
-              }
-            });
+  public void disconnectAndCloseChannel(ChannelFuture future) {
+    gracefulClose(future.channel());
+  }
+
+  public void closeChannel() {
+    gracefulClose(ctx.channel());
+  }
+
+  private static ChannelFuture gracefulClose(Channel channel) {
+    return channel
+      .disconnect()
+      .addListener(
+        disconnectFuture -> {
+          if (disconnectFuture.isSuccess()) {
+            channel
+              .close()
+              .addListener(
+                closeFuture -> {
+                  if (disconnectFuture.isSuccess()) {
+                    if (log.isTraceEnabled()) {
+                      log.trace("disconnected and closed socket {}", channel.localAddress());
+                    }
+                  } else {
+                    if (log.isWarnEnabled()) {
+                      log.warn("exception closing socket {}", channel.localAddress(),
+                        disconnectFuture.cause());
+                    }
+                  }
+                });
+          } else if (log.isWarnEnabled()) {
+            log.warn("exception disconnecting socket {}", channel.localAddress(), disconnectFuture.cause());
+          }
+        });
   }
 }