+++ ReleaseNotes.md
@@ -14,6 +14,7 @@
 
 * TGR-1313: Added support for the new VAU 'Epa für alle' format (VauEpa3).
 * TGR-1286: Tiger-Proxy: The number of open connections is now tracked and can be queried.
+* TGR-1210: Tiger-Proxy: Connection problems to remote servers are now logged.
 * TGR-1315:
   step ```TGR send {requestType} request to {tigerResolvedUrl} with contentType {string} and multiline body:``` added
 * TGR-1320: TLS-Facets now also carry the TLS-Protocol and the Cipher-Suite used for the connection